requirements.txt
asgiref==3.4.1
Django==3.2.9
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
Pillow==8.4.0
psycopg2-binary==2.9.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
transliterate==1.10.2
